extract eval scheduling logic into compute_eval_ckpt_step function
Move range check logic from inline in orchestrator.py to a proper
function in eval_utils.py so tests import the real implementation.
